:package: build: fix build scripts

+++ build.ps1
@@ -12,10 +12,13 @@ Write-Verbose -Message "move to web dir"
 Set-Location web
 Write-Verbose -Message "build web app"
 pnpm build
+if (Test-Path ../goserver/public){
 Write-Verbose -Message "clean goserver/public"
-Remove-Item ..\goserver\public\* -Recurse
+Remove-Item ..\goserver\public\* -Recurse 
+}
+
 Write-Verbose -Message "copy web app"
-Copy-Item  -Force -Path .\dist\*  -Destination ..\goserver\public\ -Recurse
+Copy-Item  -Force -Path .\dist\  -Destination ..\goserver\public\ -Recurse
 Write-Verbose -Message "move to goserver"
 Set-Location ../goserver
 Write-Verbose -Message "build goserver"
